pyproject.toml
[tool.poetry]
name = "coagulate"
version = "0.1.0"
description = ""
authors = ["Stephen Nicholas Swatman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.3.0"
hkb-diamondsquare = "^1.0.2"
colour = "^0.1.5"

[tool.poetry.group.dev.dependencies]
seaborn = "^0.12.2"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
